create separate test files for shared tests and import them as per cpython
