+++ test/unit3/test_list.py
@@ -1,1123 +1,16 @@
-"""
-Tests common to tuple, list and UserList.UserList
-"""
-
-import unittest
-import sys
-# import pickle
-# from test import support
-# from test.support import ALWAYS_EQ, NEVER_EQ
-
-class _ALWAYS_EQ:
-    """
-    Object that is equal to anything.
-    """
-    def __eq__(self, other):
-        return True
-    def __ne__(self, other):
-        return False
-
-ALWAYS_EQ = _ALWAYS_EQ()
-
-class _NEVER_EQ:
-    """
-    Object that is not equal to anything.
-    """
-    def __eq__(self, other):
-        return False
-    def __ne__(self, other):
-        return True
-    def __hash__(self):
-        return 1
-
-NEVER_EQ = _NEVER_EQ()
-
-# Various iterables
-# This is used for checking the constructor (here and in test_deque.py)
-def iterfunc(seqn):
-    'Regular generator'
-    for i in seqn:
-        yield i
-
-class Sequence:
-    'Sequence using __getitem__'
-    def __init__(self, seqn):
-        self.seqn = seqn
-    def __getitem__(self, i):
-        return self.seqn[i]
-
-class IterFunc:
-    'Sequence using iterator protocol'
-    def __init__(self, seqn):
-        self.seqn = seqn
-        self.i = 0
-    def __iter__(self):
-        return self
-    def __next__(self):
-        if self.i >= len(self.seqn): raise StopIteration
-        v = self.seqn[self.i]
-        self.i += 1
-        return v
-
-class IterGen:
-    'Sequence using iterator protocol defined with a generator'
-    def __init__(self, seqn):
-        self.seqn = seqn
-        self.i = 0
-    def __iter__(self):
-        for val in self.seqn:
-            yield val
-
-class IterNextOnly:
-    'Missing __getitem__ and __iter__'
-    def __init__(self, seqn):
-        self.seqn = seqn
-        self.i = 0
-    def __next__(self):
-        if self.i >= len(self.seqn): raise StopIteration
-        v = self.seqn[self.i]
-        self.i += 1
-        return v
-
-class IterNoNext:
-    'Iterator missing __next__()'
-    def __init__(self, seqn):
-        self.seqn = seqn
-        self.i = 0
-    def __iter__(self):
-        return self
-
-class IterGenExc:
-    'Test propagation of exceptions'
-    def __init__(self, seqn):
-        self.seqn = seqn
-        self.i = 0
-    def __iter__(self):
-        return self
-    def __next__(self):
-        3 // 0
-
-class IterFuncStop:
-    'Test immediate stop'
-    def __init__(self, seqn):
-        pass
-    def __iter__(self):
-        return self
-    def __next__(self):
-        raise StopIteration
-
-from itertools import chain
-def itermulti(seqn):
-    'Test multiple tiers of iterators'
-    return chain(map(lambda x:x, iterfunc(IterGen(Sequence(seqn)))))
-
-class LyingTuple(tuple):
-    def __iter__(self):
-        yield 1
-
-class LyingList(list):
-    def __iter__(self):
-        yield 1
-
-class CommonTestSeq:
-    # The type to be tested
-    type2test = None
-
-    def test_constructors(self):
-        l0 = []
-        l1 = [0]
-        l2 = [0, 1]
-
-        u = self.type2test()
-        u0 = self.type2test(l0)
-        u1 = self.type2test(l1)
-        u2 = self.type2test(l2)
-
-        uu = self.type2test(u)
-        uu0 = self.type2test(u0)
-        uu1 = self.type2test(u1)
-        uu2 = self.type2test(u2)
-
-        v = self.type2test(tuple(u))
-        class OtherSeq:
-            def __init__(self, initseq):
-                self.__data = initseq
-            def __len__(self):
-                return len(self.__data)
-            def __getitem__(self, i):
-                return self.__data[i]
-        s = OtherSeq(u0)
-        v0 = self.type2test(s)
-        self.assertEqual(len(v0), len(s))
-
-        s = "this is also a sequence"
-        vv = self.type2test(s)
-        self.assertEqual(len(vv), len(s))
-
-        # Create from various iteratables
-        for s in ("123", "", range(1000), ('do', 1.2), range(2000,2200,5)):
-            for g in (Sequence, IterFunc, IterGen,
-                      itermulti, iterfunc):
-                self.assertEqual(self.type2test(g(s)), self.type2test(s))
-            self.assertEqual(self.type2test(IterFuncStop(s)), self.type2test())
-            self.assertEqual(self.type2test(c for c in "123"), self.type2test("123"))
-            self.assertRaises(TypeError, self.type2test, IterNextOnly(s))
-            self.assertRaises(TypeError, self.type2test, IterNoNext(s))
-            self.assertRaises(ZeroDivisionError, self.type2test, IterGenExc(s))
-
-        # Issue #23757
-        self.assertEqual(self.type2test(LyingTuple((2,))), self.type2test((1,)))
-        self.assertEqual(self.type2test(LyingList([2])), self.type2test([1]))
-
-    def test_truth(self):
-        self.assertFalse(self.type2test())
-        self.assertTrue(self.type2test([42]))
-
-    def test_getitem(self):
-        u = self.type2test([0, 1, 2, 3, 4])
-        for i in range(len(u)):
-            self.assertEqual(u[i], i)
-            self.assertEqual(u[int(i)], i)
-        for i in range(-len(u), -1):
-            self.assertEqual(u[i], len(u)+i)
-            self.assertEqual(u[int(i)], len(u)+i)
-        self.assertRaises(IndexError, u.__getitem__, -len(u)-1)
-        self.assertRaises(IndexError, u.__getitem__, len(u))
-        self.assertRaises(ValueError, u.__getitem__, slice(0,10,0))
-
-        u = self.type2test()
-        self.assertRaises(IndexError, u.__getitem__, 0)
-        self.assertRaises(IndexError, u.__getitem__, -1)
-
-        self.assertRaises(TypeError, u.__getitem__)
-
-        a = self.type2test([10, 11])
-        self.assertEqual(a[0], 10)
-        self.assertEqual(a[1], 11)
-        self.assertEqual(a[-2], 10)
-        self.assertEqual(a[-1], 11)
-        self.assertRaises(IndexError, a.__getitem__, -3)
-        self.assertRaises(IndexError, a.__getitem__, 3)
-
-    def test_getslice(self):
-        l = [0, 1, 2, 3, 4]
-        u = self.type2test(l)
-
-        self.assertEqual(u[0:0], self.type2test())
-        self.assertEqual(u[1:2], self.type2test([1]))
-        self.assertEqual(u[-2:-1], self.type2test([3]))
-        self.assertEqual(u[-1000:1000], u)
-        self.assertEqual(u[1000:-1000], self.type2test([]))
-        self.assertEqual(u[:], u)
-        self.assertEqual(u[1:None], self.type2test([1, 2, 3, 4]))
-        self.assertEqual(u[None:3], self.type2test([0, 1, 2]))
-
-        # Extended slices
-        self.assertEqual(u[::], u)
-        self.assertEqual(u[::2], self.type2test([0, 2, 4]))
-        self.assertEqual(u[1::2], self.type2test([1, 3]))
-        self.assertEqual(u[::-1], self.type2test([4, 3, 2, 1, 0]))
-        self.assertEqual(u[::-2], self.type2test([4, 2, 0]))
-        self.assertEqual(u[3::-2], self.type2test([3, 1]))
-        self.assertEqual(u[3:3:-2], self.type2test([]))
-        self.assertEqual(u[3:2:-2], self.type2test([3]))
-        self.assertEqual(u[3:1:-2], self.type2test([3]))
-        self.assertEqual(u[3:0:-2], self.type2test([3, 1]))
-        self.assertEqual(u[::-100], self.type2test([4]))
-        self.assertEqual(u[100:-100:], self.type2test([]))
-        self.assertEqual(u[-100:100:], u)
-        self.assertEqual(u[100:-100:-1], u[::-1])
-        self.assertEqual(u[-100:100:-1], self.type2test([]))
-        self.assertEqual(u[-100:100:2], self.type2test([0, 2, 4]))
-
-        # Test extreme cases with long ints
-        a = self.type2test([0,1,2,3,4])
-        self.assertEqual(a[ -pow(2,128): 3 ], self.type2test([0,1,2]))
-        self.assertEqual(a[ 3: pow(2,145) ], self.type2test([3,4]))
-        self.assertEqual(a[3::sys.maxsize], self.type2test([3]))
-
-    def test_contains(self):
-        u = self.type2test([0, 1, 2])
-        for i in u:
-            self.assertIn(i, u)
-        for i in min(u)-1, max(u)+1:
-            self.assertNotIn(i, u)
-
-        self.assertRaises(TypeError, u.__contains__)
-
-    def test_contains_fake(self):
-        # Sequences must use rich comparison against each item
-        # (unless "is" is true, or an earlier item answered)
-        # So ALWAYS_EQ must be found in all non-empty sequences.
-        self.assertNotIn(ALWAYS_EQ, self.type2test([]))
-        self.assertIn(ALWAYS_EQ, self.type2test([1]))
-        self.assertIn(1, self.type2test([ALWAYS_EQ]))
-        self.assertNotIn(NEVER_EQ, self.type2test([]))
-        self.assertNotIn(ALWAYS_EQ, self.type2test([NEVER_EQ]))
-        self.assertIn(NEVER_EQ, self.type2test([ALWAYS_EQ]))
-
-    def test_contains_order(self):
-        # Sequences must test in-order.  If a rich comparison has side
-        # effects, these will be visible to tests against later members.
-        # In this test, the "side effect" is a short-circuiting raise.
-        class DoNotTestEq(Exception):
-            pass
-        class StopCompares:
-            def __eq__(self, other):
-                raise DoNotTestEq
-
-        checkfirst = self.type2test([1, StopCompares()])
-        self.assertIn(1, checkfirst)
-        checklast = self.type2test([StopCompares(), 1])
-        self.assertRaises(DoNotTestEq, checklast.__contains__, 1)
-
-    def test_len(self):
-        self.assertEqual(len(self.type2test()), 0)
-        self.assertEqual(len(self.type2test([])), 0)
-        self.assertEqual(len(self.type2test([0])), 1)
-        self.assertEqual(len(self.type2test([0, 1, 2])), 3)
-
-    def test_minmax(self):
-        u = self.type2test([0, 1, 2])
-        self.assertEqual(min(u), 0)
-        self.assertEqual(max(u), 2)
-
-    def test_addmul(self):
-        u1 = self.type2test([0])
-        u2 = self.type2test([0, 1])
-        self.assertEqual(u1, u1 + self.type2test())
-        self.assertEqual(u1, self.type2test() + u1)
-        self.assertEqual(u1 + self.type2test([1]), u2)
-        self.assertEqual(self.type2test([-1]) + u1, self.type2test([-1, 0]))
-        self.assertEqual(self.type2test(), u2*0)
-        self.assertEqual(self.type2test(), 0*u2)
-        self.assertEqual(self.type2test(), u2*0)
-        self.assertEqual(self.type2test(), 0*u2)
-        self.assertEqual(u2, u2*1)
-        self.assertEqual(u2, 1*u2)
-        self.assertEqual(u2, u2*1)
-        self.assertEqual(u2, 1*u2)
-        self.assertEqual(u2+u2, u2*2)
-        self.assertEqual(u2+u2, 2*u2)
-        self.assertEqual(u2+u2, u2*2)
-        self.assertEqual(u2+u2, 2*u2)
-        self.assertEqual(u2+u2+u2, u2*3)
-        self.assertEqual(u2+u2+u2, 3*u2)
-
-        class subclass(self.type2test):
-            pass
-        u3 = subclass([0, 1])
-        self.assertEqual(u3, u3*1)
-        self.assertIsNot(u3, u3*1)
-
-    def test_iadd(self):
-        u = self.type2test([0, 1])
-        u += self.type2test()
-        self.assertEqual(u, self.type2test([0, 1]))
-        u += self.type2test([2, 3])
-        self.assertEqual(u, self.type2test([0, 1, 2, 3]))
-        u += self.type2test([4, 5])
-        self.assertEqual(u, self.type2test([0, 1, 2, 3, 4, 5]))
-
-        u = self.type2test("spam")
-        u += self.type2test("eggs")
-        self.assertEqual(u, self.type2test("spameggs"))
-
-    def test_imul(self):
-        u = self.type2test([0, 1])
-        u *= 3
-        self.assertEqual(u, self.type2test([0, 1, 0, 1, 0, 1]))
-        u *= 0
-        self.assertEqual(u, self.type2test([]))
-
-    def test_getitemoverwriteiter(self):
-        # Verify that __getitem__ overrides are not recognized by __iter__
-        class T(self.type2test):
-            def __getitem__(self, key):
-                return str(key) + '!!!'
-        self.assertEqual(next(iter(T((1,2)))), 1)
-
-    def test_repeat(self):
-        for m in range(4):
-            s = tuple(range(m))
-            for n in range(-3, 5):
-                self.assertEqual(self.type2test(s*n), self.type2test(s)*n)
-            self.assertEqual(self.type2test(s)*(-4), self.type2test([]))
-            self.assertEqual(id(s), id(s*1))
-
-    def test_bigrepeat(self):
-        if sys.maxsize <= 2147483647:
-            x = self.type2test([0])
-            x *= 2**16
-            self.assertRaises(MemoryError, x.__mul__, 2**16)
-            if hasattr(x, '__imul__'):
-                self.assertRaises(MemoryError, x.__imul__, 2**16)
-
-    def test_subscript(self):
-        a = self.type2test([10, 11])
-        self.assertEqual(a.__getitem__(0), 10)
-        self.assertEqual(a.__getitem__(1), 11)
-        self.assertEqual(a.__getitem__(-2), 10)
-        self.assertEqual(a.__getitem__(-1), 11)
-        self.assertRaises(IndexError, a.__getitem__, -3)
-        self.assertRaises(IndexError, a.__getitem__, 3)
-        self.assertEqual(a.__getitem__(slice(0,1)), self.type2test([10]))
-        self.assertEqual(a.__getitem__(slice(1,2)), self.type2test([11]))
-        self.assertEqual(a.__getitem__(slice(0,2)), self.type2test([10, 11]))
-        self.assertEqual(a.__getitem__(slice(0,3)), self.type2test([10, 11]))
-        self.assertEqual(a.__getitem__(slice(3,5)), self.type2test([]))
-        self.assertRaises(ValueError, a.__getitem__, slice(0, 10, 0))
-        self.assertRaises(TypeError, a.__getitem__, 'x')
-
-    def test_count(self):
-        a = self.type2test([0, 1, 2])*3
-        self.assertEqual(a.count(0), 3)
-        self.assertEqual(a.count(1), 3)
-        self.assertEqual(a.count(3), 0)
-
-        self.assertEqual(a.count(ALWAYS_EQ), 9)
-        self.assertEqual(self.type2test([ALWAYS_EQ, ALWAYS_EQ]).count(1), 2)
-        self.assertEqual(self.type2test([ALWAYS_EQ, ALWAYS_EQ]).count(NEVER_EQ), 2)
-        self.assertEqual(self.type2test([NEVER_EQ, NEVER_EQ]).count(ALWAYS_EQ), 0)
-
-        self.assertRaises(TypeError, a.count)
-
-        class BadExc(Exception):
-            pass
-
-        class BadCmp:
-            def __eq__(self, other):
-                if other == 2:
-                    raise BadExc()
-                return False
-
-        self.assertRaises(BadExc, a.count, BadCmp())
-
-    def test_index(self):
-        u = self.type2test([0, 1])
-        self.assertEqual(u.index(0), 0)
-        self.assertEqual(u.index(1), 1)
-        self.assertRaises(ValueError, u.index, 2)
-
-        u = self.type2test([-2, -1, 0, 0, 1, 2])
-        self.assertEqual(u.count(0), 2)
-        self.assertEqual(u.index(0), 2)
-        self.assertEqual(u.index(0, 2), 2)
-        self.assertEqual(u.index(-2, -10), 0)
-        self.assertEqual(u.index(0, 3), 3)
-        self.assertEqual(u.index(0, 3, 4), 3)
-        self.assertRaises(ValueError, u.index, 2, 0, -10)
-
-        self.assertEqual(u.index(ALWAYS_EQ), 0)
-        self.assertEqual(self.type2test([ALWAYS_EQ, ALWAYS_EQ]).index(1), 0)
-        self.assertEqual(self.type2test([ALWAYS_EQ, ALWAYS_EQ]).index(NEVER_EQ), 0)
-        self.assertRaises(ValueError, self.type2test([NEVER_EQ, NEVER_EQ]).index, ALWAYS_EQ)
-
-        self.assertRaises(TypeError, u.index)
-
-        class BadExc(Exception):
-            pass
-
-        class BadCmp:
-            def __eq__(self, other):
-                if other == 2:
-                    raise BadExc()
-                return False
-
-        a = self.type2test([0, 1, 2, 3])
-        self.assertRaises(BadExc, a.index, BadCmp())
-
-        a = self.type2test([-2, -1, 0, 0, 1, 2])
-        self.assertEqual(a.index(0), 2)
-        self.assertEqual(a.index(0, 2), 2)
-        self.assertEqual(a.index(0, -4), 2)
-        self.assertEqual(a.index(-2, -10), 0)
-        self.assertEqual(a.index(0, 3), 3)
-        self.assertEqual(a.index(0, -3), 3)
-        self.assertEqual(a.index(0, 3, 4), 3)
-        self.assertEqual(a.index(0, -3, -2), 3)
-        self.assertEqual(a.index(0, -4*sys.maxsize, 4*sys.maxsize), 2)
-        self.assertRaises(ValueError, a.index, 0, 4*sys.maxsize,-4*sys.maxsize)
-        self.assertRaises(ValueError, a.index, 2, 0, -10)
-
-    # def test_pickle(self):
-    #     lst = self.type2test([4, 5, 6, 7])
-    #     for proto in range(pickle.HIGHEST_PROTOCOL + 1):
-    #         lst2 = pickle.loads(pickle.dumps(lst, proto))
-    #         self.assertEqual(lst2, lst)
-    #         self.assertNotEqual(id(lst2), id(lst))
-
-    # def test_free_after_iterating(self):
-    #     support.check_free_after_iterating(self, iter, self.type2test)
-    #     support.check_free_after_iterating(self, reversed, self.type2test)
-
-
-
-"""
-Tests common to list and UserList.UserList
-"""
-
-import sys
-# import os
-# from functools import cmp_to_key
-def cmp_to_key(mycmp):
-    """Convert a cmp= function into a key= function"""
-    class K(object):
-        __slots__ = ['obj']
-        def __init__(self, obj):
-            self.obj = obj
-        def __lt__(self, other):
-            return mycmp(self.obj, other.obj) < 0
-        def __gt__(self, other):
-            return mycmp(self.obj, other.obj) > 0
-        def __eq__(self, other):
-            return mycmp(self.obj, other.obj) == 0
-        def __le__(self, other):
-            return mycmp(self.obj, other.obj) <= 0
-        def __ge__(self, other):
-            return mycmp(self.obj, other.obj) >= 0
-        __hash__ = None
-    return K
-
-# from test import support, seq_tests
-# from test.support import ALWAYS_EQ, NEVER_EQ
-
-
-class CommonTestList(CommonTestSeq):
-
-    def test_init(self):
-        # Iterable arg is optional
-        self.assertEqual(self.type2test([]), self.type2test())
-
-        # Init clears previous values
-        a = self.type2test([1, 2, 3])
-        a.__init__()
-        self.assertEqual(a, self.type2test([]))
-
-        # Init overwrites previous values
-        a = self.type2test([1, 2, 3])
-        a.__init__([4, 5, 6])
-        self.assertEqual(a, self.type2test([4, 5, 6]))
-
-        # Mutables always return a new object
-        b = self.type2test(a)
-        self.assertNotEqual(id(a), id(b))
-        self.assertEqual(a, b)
-
-    def test_getitem_error(self):
-        a = []
-        msg = "list indices must be integers or slices"
-        with self.assertRaises(TypeError) as c:
-            a['a']
-        self.assertIn(msg, c.exception.args[0])
-
-    def test_setitem_error(self):
-        a = []
-        msg = "list indices must be integers or slices"
-        with self.assertRaises(TypeError) as c:
-            a['a'] = "python"
-        self.assertIn(msg, c.exception.args[0])
-
-    def test_repr(self):
-        l0 = []
-        l2 = [0, 1, 2]
-        a0 = self.type2test(l0)
-        a2 = self.type2test(l2)
-
-        self.assertEqual(str(a0), str(l0))
-        self.assertEqual(repr(a0), repr(l0))
-        self.assertEqual(repr(a2), repr(l2))
-        self.assertEqual(str(a2), "[0, 1, 2]")
-        self.assertEqual(repr(a2), "[0, 1, 2]")
-
-        a2.append(a2)
-        a2.append(3)
-        self.assertEqual(str(a2), "[0, 1, 2, [...], 3]")
-        self.assertEqual(repr(a2), "[0, 1, 2, [...], 3]")
-
-    # def test_repr_deep(self):
-    #     a = self.type2test([])
-    #     for i in range(sys.getrecursionlimit() + 100):
-    #         a = self.type2test([a])
-    #     self.assertRaises(RecursionError, repr, a)
-
-    # def test_print(self):
-    #     d = self.type2test(range(200))
-    #     d.append(d)
-    #     d.extend(range(200,400))
-    #     d.append(d)
-    #     d.append(400)
-    #     try:
-    #         with open(support.TESTFN, "w") as fo:
-    #             fo.write(str(d))
-    #         with open(support.TESTFN, "r") as fo:
-    #             self.assertEqual(fo.read(), repr(d))
-    #     finally:
-    #         os.remove(support.TESTFN)
-
-    def test_set_subscript(self):
-        a = self.type2test(range(20))
-        self.assertRaises(ValueError, a.__setitem__, slice(0, 10, 0), [1,2,3])
-        self.assertRaises(TypeError, a.__setitem__, slice(0, 10), 1)
-        self.assertRaises(ValueError, a.__setitem__, slice(0, 10, 2), [1,2])
-        self.assertRaises(TypeError, a.__getitem__, 'x', 1)
-        a[slice(2,10,3)] = [1,2,3]
-        self.assertEqual(a, self.type2test([0, 1, 1, 3, 4, 2, 6, 7, 3,
-                                            9, 10, 11, 12, 13, 14, 15,
-                                            16, 17, 18, 19]))
-
-    def test_reversed(self):
-        a = self.type2test(range(20))
-        r = reversed(a)
-        self.assertEqual(list(r), self.type2test(range(19, -1, -1)))
-        self.assertRaises(StopIteration, next, r)
-        self.assertEqual(list(reversed(self.type2test())),
-                         self.type2test())
-        # Bug 3689: make sure list-reversed-iterator doesn't have __len__
-        self.assertRaises(TypeError, len, reversed([1,2,3]))
-
-    def test_setitem(self):
-        a = self.type2test([0, 1])
-        a[0] = 0
-        a[1] = 100
-        self.assertEqual(a, self.type2test([0, 100]))
-        a[-1] = 200
-        self.assertEqual(a, self.type2test([0, 200]))
-        a[-2] = 100
-        self.assertEqual(a, self.type2test([100, 200]))
-        self.assertRaises(IndexError, a.__setitem__, -3, 200)
-        self.assertRaises(IndexError, a.__setitem__, 2, 200)
-
-        a = self.type2test([])
-        self.assertRaises(IndexError, a.__setitem__, 0, 200)
-        self.assertRaises(IndexError, a.__setitem__, -1, 200)
-        self.assertRaises(TypeError, a.__setitem__)
-
-        a = self.type2test([0,1,2,3,4])
-        a[0] = 1
-        a[1] = 2
-        a[2] = 3
-        self.assertEqual(a, self.type2test([1,2,3,3,4]))
-        a[0] = 5
-        a[1] = 6
-        a[2] = 7
-        self.assertEqual(a, self.type2test([5,6,7,3,4]))
-        a[-2] = 88
-        a[-1] = 99
-        self.assertEqual(a, self.type2test([5,6,7,88,99]))
-        a[-2] = 8
-        a[-1] = 9
-        self.assertEqual(a, self.type2test([5,6,7,8,9]))
-
-        msg = "list indices must be integers or slices"
-        with self.assertRaises(TypeError) as c:
-            a['a'] = "python"
-        self.assertIn(msg, c.exception.args[0])
-
-    def test_delitem(self):
-        a = self.type2test([0, 1])
-        del a[1]
-        self.assertEqual(a, [0])
-        del a[0]
-        self.assertEqual(a, [])
-
-        a = self.type2test([0, 1])
-        del a[-2]
-        self.assertEqual(a, [1])
-        del a[-1]
-        self.assertEqual(a, [])
-
-        a = self.type2test([0, 1])
-        self.assertRaises(IndexError, a.__delitem__, -3)
-        self.assertRaises(IndexError, a.__delitem__, 2)
-
-        a = self.type2test([])
-        self.assertRaises(IndexError, a.__delitem__, 0)
-
-        self.assertRaises(TypeError, a.__delitem__)
-
-    def test_setslice(self):
-        l = [0, 1]
-        a = self.type2test(l)
-
-        for i in range(-3, 4):
-            a[:i] = l[:i]
-            self.assertEqual(a, l)
-            a2 = a[:]
-            a2[:i] = a[:i]
-            self.assertEqual(a2, a)
-            a[i:] = l[i:]
-            self.assertEqual(a, l)
-            a2 = a[:]
-            a2[i:] = a[i:]
-            self.assertEqual(a2, a)
-            for j in range(-3, 4):
-                a[i:j] = l[i:j]
-                self.assertEqual(a, l)
-                a2 = a[:]
-                a2[i:j] = a[i:j]
-                self.assertEqual(a2, a)
-
-        aa2 = a2[:]
-        aa2[:0] = [-2, -1]
-        self.assertEqual(aa2, [-2, -1, 0, 1])
-        aa2[0:] = []
-        self.assertEqual(aa2, [])
-
-        a = self.type2test([1, 2, 3, 4, 5])
-        a[:-1] = a
-        self.assertEqual(a, self.type2test([1, 2, 3, 4, 5, 5]))
-        a = self.type2test([1, 2, 3, 4, 5])
-        a[1:] = a
-        self.assertEqual(a, self.type2test([1, 1, 2, 3, 4, 5]))
-        a = self.type2test([1, 2, 3, 4, 5])
-        a[1:-1] = a
-        self.assertEqual(a, self.type2test([1, 1, 2, 3, 4, 5, 5]))
-
-        a = self.type2test([])
-        a[:] = tuple(range(10))
-        self.assertEqual(a, self.type2test(range(10)))
-
-        self.assertRaises(TypeError, a.__setitem__, slice(0, 1, 5))
-
-        self.assertRaises(TypeError, a.__setitem__)
-
-    def test_delslice(self):
-        a = self.type2test([0, 1])
-        del a[1:2]
-        del a[0:1]
-        self.assertEqual(a, self.type2test([]))
-
-        a = self.type2test([0, 1])
-        del a[1:2]
-        del a[0:1]
-        self.assertEqual(a, self.type2test([]))
-
-        a = self.type2test([0, 1])
-        del a[-2:-1]
-        self.assertEqual(a, self.type2test([1]))
-
-        a = self.type2test([0, 1])
-        del a[-2:-1]
-        self.assertEqual(a, self.type2test([1]))
-
-        a = self.type2test([0, 1])
-        del a[1:]
-        del a[:1]
-        self.assertEqual(a, self.type2test([]))
-
-        a = self.type2test([0, 1])
-        del a[1:]
-        del a[:1]
-        self.assertEqual(a, self.type2test([]))
-
-        a = self.type2test([0, 1])
-        del a[-1:]
-        self.assertEqual(a, self.type2test([0]))
-
-        a = self.type2test([0, 1])
-        del a[-1:]
-        self.assertEqual(a, self.type2test([0]))
-
-        a = self.type2test([0, 1])
-        del a[:]
-        self.assertEqual(a, self.type2test([]))
-
-    def test_append(self):
-        a = self.type2test([])
-        a.append(0)
-        a.append(1)
-        a.append(2)
-        self.assertEqual(a, self.type2test([0, 1, 2]))
-
-        self.assertRaises(TypeError, a.append)
-
-    def test_extend(self):
-        a1 = self.type2test([0])
-        a2 = self.type2test((0, 1))
-        a = a1[:]
-        a.extend(a2)
-        self.assertEqual(a, a1 + a2)
-
-        a.extend(self.type2test([]))
-        self.assertEqual(a, a1 + a2)
-
-        a.extend(a)
-        self.assertEqual(a, self.type2test([0, 0, 1, 0, 0, 1]))
-
-        a = self.type2test("spam")
-        a.extend("eggs")
-        self.assertEqual(a, list("spameggs"))
-
-        l1 = [42]
-        l2 = l1
-        l1 += [99]
-        self.assertEqual(l1, l2)
-        self.assertEqual(l1, [42, 99])
-        l1 += l1
-        self.assertEqual(l1, [42, 99, 42, 99])
-
-        self.assertRaises(TypeError, a.extend, None)
-        m = [[1,2,3],2,3]
-        m.extend(m)
-        self.assertEqual(m, [[1, 2, 3], 2, 3, [1, 2, 3], 2, 3])
-        self.assertRaises(TypeError, lambda x: x + 1, [1])
-
-
-        self.assertRaises(TypeError, a.extend, None)
-        self.assertRaises(TypeError, a.extend)
-
-        # overflow test. issue1621
-        class CustomIter:
-            def __iter__(self):
-                return self
-            def __next__(self):
-                raise StopIteration
-            def __length_hint__(self):
-                return sys.maxsize
-        a = self.type2test([1,2,3,4])
-        a.extend(CustomIter())
-        self.assertEqual(a, [1,2,3,4])
-
-
-    def test_insert(self):
-        a = self.type2test([0, 1, 2])
-        a.insert(0, -2)
-        a.insert(1, -1)
-        a.insert(2, 0)
-        self.assertEqual(a, [-2, -1, 0, 0, 1, 2])
-
-        b = a[:]
-        b.insert(-2, "foo")
-        b.insert(-200, "left")
-        b.insert(200, "right")
-        self.assertEqual(b, self.type2test(["left",-2,-1,0,0,"foo",1,2,"right"]))
-
-        self.assertRaises(TypeError, a.insert)
-        things = ['hi', 'a', 'b', 'c']
-        things.insert(len(things), 'bye')
-        self.assertEqual(things, ['hi', 'a', 'b', 'c', 'bye'])
-        things.insert(len(things)+3, 'surpise')
-        self.assertEqual(things, ['hi', 'a', 'b', 'c', 'bye', 'surpise'])
-
-    def test_pop(self):
-        a = self.type2test([-1, 0, 1])
-        a.pop()
-        self.assertEqual(a, [-1, 0])
-        a.pop(0)
-        self.assertEqual(a, [0])
-        self.assertRaises(IndexError, a.pop, 5)
-        a.pop(0)
-        self.assertEqual(a, [])
-        self.assertRaises(IndexError, a.pop)
-        self.assertRaises(TypeError, a.pop, 42, 42)
-        a = self.type2test([0, 10, 20, 30, 40])
-
-    def test_remove(self):
-        a = self.type2test([0, 0, 1])
-        a.remove(1)
-        self.assertEqual(a, [0, 0])
-        a.remove(0)
-        self.assertEqual(a, [0])
-        a.remove(0)
-        self.assertEqual(a, [])
-
-        self.assertRaises(ValueError, a.remove, 0)
-
-        self.assertRaises(TypeError, a.remove)
-
-        a = self.type2test([1, 2])
-        self.assertRaises(ValueError, a.remove, NEVER_EQ)
-        self.assertEqual(a, [1, 2])
-        a.remove(ALWAYS_EQ)
-        self.assertEqual(a, [2])
-        a = self.type2test([ALWAYS_EQ])
-        a.remove(1)
-        self.assertEqual(a, [])
-        a = self.type2test([ALWAYS_EQ])
-        a.remove(NEVER_EQ)
-        self.assertEqual(a, [])
-        a = self.type2test([NEVER_EQ])
-        self.assertRaises(ValueError, a.remove, ALWAYS_EQ)
-
-        class BadExc(Exception):
-            pass
-
-        class BadCmp:
-            def __eq__(self, other):
-                if other == 2:
-                    raise BadExc()
-                return False
-
-        a = self.type2test([0, 1, 2, 3])
-        self.assertRaises(BadExc, a.remove, BadCmp())
-
-        class BadCmp2:
-            def __eq__(self, other):
-                raise BadExc()
-
-        d = self.type2test('abcdefghcij')
-        d.remove('c')
-        self.assertEqual(d, self.type2test('abdefghcij'))
-        d.remove('c')
-        self.assertEqual(d, self.type2test('abdefghij'))
-        self.assertRaises(ValueError, d.remove, 'c')
-        self.assertEqual(d, self.type2test('abdefghij'))
-
-        # Handle comparison errors
-        d = self.type2test(['a', 'b', BadCmp2(), 'c'])
-        e = self.type2test(d)
-        self.assertRaises(BadExc, d.remove, 'c')
-        for x, y in zip(d, e):
-            # verify that original order and values are retained.
-            self.assertIs(x, y)
-
-    def test_index(self):
-        super().test_index()
-        a = self.type2test([-2, -1, 0, 0, 1, 2])
-        a.remove(0)
-        self.assertRaises(ValueError, a.index, 2, 0, 4)
-        self.assertEqual(a, self.type2test([-2, -1, 0, 1, 2]))
-
-        # Test modifying the list during index's iteration
-        class EvilCmp:
-            def __init__(self, victim):
-                self.victim = victim
-            def __eq__(self, other):
-                del self.victim[:]
-                return False
-        a = self.type2test()
-        a[:] = [EvilCmp(a) for _ in range(100)]
-        # This used to seg fault before patch #1005778
-        self.assertRaises(ValueError, a.index, None)
-
-    def test_reverse(self):
-        u = self.type2test([-2, -1, 0, 1, 2])
-        u2 = u[:]
-        u.reverse()
-        self.assertEqual(u, [2, 1, 0, -1, -2])
-        u.reverse()
-        self.assertEqual(u, u2)
-
-        self.assertRaises(TypeError, u.reverse, 42)
-
-    def test_shallow_copy(self):
-        a = [0,[1]]
-        b = a.copy()
-        self.assertEqual(a,b)
-        a[1][0] = 2
-        self.assertEqual(a,b)
-        a[0] = 5
-        self.assertNotEqual(a,b)
-        l = [1,[2],3]
-        m = list(l)
-        self.assertEqual(l,m)
-        l[1][0] = 3
-        self.assertEqual(l,m)
-        m[0] = 10
-        self.assertNotEqual(l,m)
-        
-    def test_clear(self):
-        u = self.type2test([2, 3, 4])
-        u.clear()
-        self.assertEqual(u, [])
-
-        u = self.type2test([])
-        u.clear()
-        self.assertEqual(u, [])
-
-        u = self.type2test([])
-        u.append(1)
-        u.clear()
-        u.append(2)
-        self.assertEqual(u, [2])
-
-        self.assertRaises(TypeError, u.clear, None)
-
-    def test_copy(self):
-        u = self.type2test([1, 2, 3])
-        v = u.copy()
-        self.assertEqual(v, [1, 2, 3])
-
-        u = self.type2test([])
-        v = u.copy()
-        self.assertEqual(v, [])
-
-        # test that it's indeed a copy and not a reference
-        u = self.type2test(['a', 'b'])
-        v = u.copy()
-        v.append('i')
-        self.assertEqual(u, ['a', 'b'])
-        self.assertEqual(v, u + ['i'])
-
-        # test that it's a shallow, not a deep copy
-        u = self.type2test([1, 2, [3, 4], 5])
-        v = u.copy()
-        self.assertEqual(u, v)
-        self.assertIs(v[3], u[3])
-
-        self.assertRaises(TypeError, u.copy, None)
-
-    def test_sort(self):
-        u = self.type2test([1, 0])
-        u.sort()
-        self.assertEqual(u, [0, 1])
-
-        u = self.type2test([2,1,0,-1,-2])
-        u.sort()
-        self.assertEqual(u, self.type2test([-2,-1,0,1,2]))
-
-        self.assertRaises(TypeError, u.sort, 42, 42)
-
-        def revcmp(a, b):
-            if a == b:
-                return 0
-            elif a < b:
-                return 1
-            else: # a > b
-                return -1
-        u.sort(key=cmp_to_key(revcmp))
-        self.assertEqual(u, self.type2test([2,1,0,-1,-2]))
-
-        # The following dumps core in unpatched Python 1.5:
-        def myComparison(x,y):
-            xmod, ymod = x%3, y%7
-            if xmod == ymod:
-                return 0
-            elif xmod < ymod:
-                return -1
-            else: # xmod > ymod
-                return 1
-        z = self.type2test(range(12))
-        z.sort(key=cmp_to_key(myComparison))
-
-        self.assertRaises(TypeError, z.sort, 2)
-
-        def selfmodifyingComparison(x,y):
-            z.append(1)
-            if x == y:
-                return 0
-            elif x < y:
-                return -1
-            else: # x > y
-                return 1
-        self.assertRaises(ValueError, z.sort,
-                          key=cmp_to_key(selfmodifyingComparison))
-
-        self.assertRaises(TypeError, z.sort, 42, 42, 42, 42)
-
-    def test_slice(self):
-        u = self.type2test("spam")
-        u[:2] = "h"
-        self.assertEqual(u, list("ham"))
-
-    def test_iadd(self):
-        super().test_iadd()
-        u = self.type2test([0, 1])
-        u2 = u
-        u += [2, 3]
-        self.assertIs(u, u2)
-
-        u = self.type2test("spam")
-        u += "eggs"
-        self.assertEqual(u, self.type2test("spameggs"))
-
-        self.assertRaises(TypeError, u.__iadd__, None)
-
-    def test_imul(self):
-        super().test_imul()
-        s = self.type2test([])
-        oldid = id(s)
-        s *= 10
-        self.assertEqual(id(s), oldid)
-
-    def test_extendedslicing(self):
-        #  subscript
-        a = self.type2test([0,1,2,3,4])
-
-        #  deletion
-        del a[::2]
-        self.assertEqual(a, self.type2test([1,3]))
-        a = self.type2test(range(5))
-        del a[1::2]
-        self.assertEqual(a, self.type2test([0,2,4]))
-        a = self.type2test(range(5))
-        del a[1::-2]
-        self.assertEqual(a, self.type2test([0,2,3,4]))
-        a = self.type2test(range(10))
-        del a[::1000]
-        self.assertEqual(a, self.type2test([1, 2, 3, 4, 5, 6, 7, 8, 9]))
-        #  assignment
-        a = self.type2test(range(10))
-        a[::2] = [-1]*5
-        self.assertEqual(a, self.type2test([-1, 1, -1, 3, -1, 5, -1, 7, -1, 9]))
-        a = self.type2test(range(10))
-        a[::-4] = [10]*3
-        self.assertEqual(a, self.type2test([0, 10, 2, 3, 4, 10, 6, 7, 8 ,10]))
-        a = self.type2test(range(4))
-        a[::-1] = a
-        self.assertEqual(a, self.type2test([3, 2, 1, 0]))
-        a = self.type2test(range(10))
-        b = a[:]
-        c = a[:]
-        a[2:3] = self.type2test(["two", "elements"])
-        b[slice(2,3)] = self.type2test(["two", "elements"])
-        c[2:3:] = self.type2test(["two", "elements"])
-        self.assertEqual(a, b)
-        self.assertEqual(a, c)
-        a = self.type2test(range(10))
-        a[::2] = tuple(range(5))
-        self.assertEqual(a, self.type2test([0, 1, 1, 3, 2, 5, 3, 7, 4, 9]))
-        # test issue7788
-        a = self.type2test(range(10))
-        del a[9::1<<333]
-        x = [1,2,3,4,5]
-        y = x[::-1]
-        self.assertEqual(y, [5, 4, 3, 2, 1])
-        l = [0,1,2,3,4]
-        self.assertTrue(l[0:3] == l[:3] == l[None:3] == [0,1,2])
-        self.assertTrue(l[0:] == l[0:None] == l[:] == [0,1,2,3,4])
-        l = [0, 1, 2, 3]
-        error1, error2, = None, None
-
-        def foo(x, y):
-            return l[x: : y]
-        self.assertRaises(ValueError, foo, 1, 0)
-        def foo2(x, y, z):
-            return l[x : y : z]
-        self.assertRaises(ValueError, foo2, 1, 3, 0)
-
-    def test_constructor_exception_handling(self):
-        # Bug #1242657
-        class F(object):
-            def __iter__(self):
-                raise AssertionError
-        self.assertRaises(AssertionError, list, F())
-
-    def test_exhausted_iterator(self):
-        a = self.type2test([1, 2, 3])
-        exhit = iter(a)
-        empit = iter(a)
-        for x in exhit:  # exhaust the iterator
-            next(empit)  # not exhausted
-        a.append(9)
-        self.assertEqual(list(exhit), [])
-        self.assertEqual(list(empit), [9])
-        self.assertEqual(a, self.type2test([1, 2, 3, 9]))
-
 
 """
 cpython test_list.py
 """
 
-
+import unittest
 import sys
+import list_tests
 # from test import list_tests
 # from test.support import cpython_only
 # import pickle
-import unittest
 
-class ListTest(CommonTestList, unittest.TestCase):
+class ListTest(list_tests.CommonTest):
     type2test = list
 
     def test_basic(self):
